delete confirmation dialog for thumbnail deletion
