+++ web/src/pages/Status/ServiceDetailsSettings/PTeamServiceImageUploadDeleteButton.jsx
@@ -16,15 +16,13 @@ import {
   serviceImageMaxSize,
 } from "../../../utils/const";
 
-import { DeleteServiceImageAlertDialog } from "./DeleteServiceImageAlertDialog";
-
 export function PTeamServiceImageUploadDeleteButton(props) {
   const { setImageFileData, setImageDeleteFlag, setImagePreview } = props;
 
   const [anchorEl, setAnchorEl] = useState(null);
-  const [isDeleteDialogOpen, setIsDeleteDialogOpen] = useState(false);
 
   const { enqueueSnackbar } = useSnackbar();
+  const serviceDetailsSetttingNoImageUrl = "images/720x480.png";
 
   const open = Boolean(anchorEl);
   const handleClick = (event) => {
@@ -65,6 +63,12 @@ export function PTeamServiceImageUploadDeleteButton(props) {
     reader.readAsDataURL(event.target.files[0]);
   };
 
+  const handleDelete = () => {
+    setImageFileData(null);
+    setImageDeleteFlag(true);
+    setImagePreview(serviceDetailsSetttingNoImageUrl);
+  };
+
   const VisuallyHiddenInput = styled("input")({
     clip: "rect(0 0 0 0)",
     clipPath: "inset(50%)",
@@ -94,7 +98,7 @@ export function PTeamServiceImageUploadDeleteButton(props) {
           <ArrowDropDownIcon />
         </Button>
       </ButtonGroup>
-      <Menu anchorEl={anchorEl} open={open} onClose={handleClose}>
+      <Menu anchorEl={anchorEl} open={open} onClose={handleClose} onClick={handleDelete}>
         <MenuItem
           onClick={() => {
             handleClose();
@@ -105,13 +109,6 @@ export function PTeamServiceImageUploadDeleteButton(props) {
           Delete image
         </MenuItem>
       </Menu>
-      <DeleteServiceImageAlertDialog
-        isDeleteDialogOpen={isDeleteDialogOpen}
-        setIsDeleteDialogOpen={setIsDeleteDialogOpen}
-        setImageFileData={setImageFileData}
-        setImageDeleteFlag={setImageDeleteFlag}
-        setImagePreview={setImagePreview}
-      />
     </>
   );
 }
